Move main API to lib.rs
