+++ dhall/src/semantics/nze/value.rs
@@ -4,7 +4,6 @@ use std::rc::Rc;
 
 use crate::error::{TypeError, TypeMessage};
 use crate::semantics::nze::visitor;
-use crate::semantics::phase::{Normalized, NormalizedExpr, ToExprOptions};
 use crate::semantics::Binder;
 use crate::semantics::{apply_any, normalize_tyexpr_whnf, normalize_whnf};
 use crate::semantics::{type_of_builtin, typecheck, TyExpr, TyExprKind};
@@ -13,6 +12,7 @@ use crate::syntax::{
     BinOp, Builtin, Const, ExprKind, Integer, InterpolatedTextContents, Label,
     NaiveDouble, Natural, Span,
 };
+use crate::{Normalized, NormalizedExpr, ToExprOptions};
 
 use self::Form::{Unevaled, WHNF};
