+++ dhall/src/semantics/tck/tyexpr.rs
@@ -1,9 +1,9 @@
 use crate::error::{TypeError, TypeMessage};
 use crate::semantics::normalize_tyexpr_whnf;
-use crate::semantics::phase::Normalized;
-use crate::semantics::phase::{NormalizedExpr, ToExprOptions};
 use crate::semantics::{rc, NameEnv, NzEnv, TyEnv, Value};
 use crate::syntax::{ExprKind, Span, V};
+use crate::Normalized;
+use crate::{NormalizedExpr, ToExprOptions};
 
 pub(crate) type Type = Value;
