+++ dhall/src/syntax/text/parser.rs
@@ -5,7 +5,6 @@ use std::rc::Rc;
 
 use pest_consume::{match_nodes, Parser};
 
-use crate::semantics::phase::Normalized;
 use crate::syntax;
 use crate::syntax::map::{DupTreeMap, DupTreeSet};
 use crate::syntax::ExprKind::*;
@@ -14,6 +13,7 @@ use crate::syntax::{
     InterpolatedText, InterpolatedTextContents, Label, NaiveDouble, Natural,
     Scheme, Span, UnspannedExpr, URL, V,
 };
+use crate::Normalized;
 
 // This file consumes the parse tree generated by pest and turns it into
 // our own AST. All those custom macros should eventually moved into
